// frontend/src/types/api.ts
export interface UserProfile {
  id: number;
  telegram_id: number;
  first_name: string;
  username: string;
  photo_url: string;
  level: number;
  energy: number;
  max_energy: number;
  spins: number;
  diamonds: number;
  balance_usd: number;
  ton_wallet: string; // Used for BEP-20 address (0x...)
  phone: string;
  goal_usd: number;   // Default $1.00
  goal_left: number;  // Remaining to cashout
  is_admin: boolean;  // True if user is authorized admin
  has_claimed_channel_reward?: boolean;
}